refactor(use data value set): extract mapDataValuesToFormInitialValues fn
